pyproject.toml
[project]
name = "k-thon-function-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "azure-functions>=1.23.0",
    "requests>=2.32.5",
    "requests-toolbelt>=1.0.0",
]
